글 작성 페이지 수정 - 글 작성시, db 저장 후 글 작성 된 화면으로 이동

+++ BookieProject/src/main/java/com/project/bookie/controller/ClubBoardController.java
@@ -80,7 +80,7 @@ public class ClubBoardController {
 	}
 
 	@GetMapping("/detail")
-	public String goDetailPage(Model m, @RequestParam(value = "b", defaultValue = "1", required = false) int boardId) {
+	public String goDetailPage(Model m, @RequestParam(value = "b", defaultValue = "1", required = false) long boardId) {
 		Board board = service.getBoardByBoardById(boardId);
 		m.addAttribute("board", board);
 		return "bookClub/bookClubBoardDetail.jsp?b=" + boardId;
